Comments and Changes
Commented properly and made changes to the weapons, ade them ints and not floats for future use with stat checks for hit/miss ratio and status effects for advanced combat

+++ Assets/Scripts/Entity/CharacterBase.cs
@@ -62,7 +62,7 @@ public class CharacterBase : MonoBehaviour
     {
         painThreshold = strong / 2;
     }
-    public void TakeDamage(float damage)
+    public void TakeDamage(int damage)
     {
         
             currentHealth -= damage;
